refactor: extract runEval to eval.ts, rename getConfig to getEvalConfig, make index.ts routing-only
